requirements-gpu.txt
tensorflow-gpu==2.3.0rc0
opencv-python==4.1.1.26
scikit-image==0.17.2
filterpy==1.4.5
pandas
numpy==1.18.5
shapely
scipy>=1.4.1
sklearn
fastdtw
networkx
seaborn
torch
h5py
numba>=0.50.0
pyDOE2>=1.3.0
lap==0.4.0
lxml
tqdm
absl-py
matplotlib
easydict
pillow
traja
smbus
pyserial
